% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdl_grid}
\alias{cdl_grid}
\title{CDL Grid}
\format{
A data frame with 3 variables: \code{location}, \code{cdl_code} and
  \code{n}
}
\usage{
cdl_grid
}
\description{
2020 CDL aggregated to 1/16th degree resolution commonly used in
hydroclimatalogy data
}
\keyword{datasets}
