% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elev_grid}
\alias{elev_grid}
\title{Elevation Grid}
\format{
A data frame with 3 variables: \code{location}, \code{min_elev} and
  \code{median_elev}
}
\usage{
elev_grid
}
\description{
CONUS dem aggregated to 1/16th degree resolution commonly used in
hydroclimatalogy data. Elevations are recorded as meters
above sea level and aggregated to min and median for each grid cell.
}
\keyword{datasets}
