% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_stats}
\alias{grid_stats}
\title{Grid Location data}
\format{
A data frame with 6 variables: \code{location}, \code{lat}, \code{lon}, \code{state}, \code{county} and
  \code{area_km2}
}
\usage{
grid_stats
}
\description{
Lat, long, state, county, and area (km^2) information
for each CONUS 1/16th degree grid
}
\keyword{datasets}
